BREAKING: ACL configuration can be queried #7330
The new `Acl::show()` returns an array that describe the current ACL
configuration for the given role. If multiple roles are given, it will
cumulate all permissions.

It is a breaking change because inheriting classes must now call
`Acl::__construct()`.

`DebugAcl` is strictly reserved for internal use. Its purpose is to
mirror the normal ACL configuration, but with assertion disabled. So we
can query which assertion is configured without actually running it.

Also, from now on all assertions must implement the new `NamedAssertion`
in order to provide a human friendly name.

Finally, the new `AbstractAcl` allows applications to easily unit test
their ACL configuration.