BREAKING First/last login are stored in user #9020
This is much faster to query, and allow to delete much more logs,
drastically reducing DB size.

We also delete pretty much all logs defined by Felix which are older
than 2 months. Because we never needed them in the past two years, and
there is no reason to believe we will need them in the future.