Searching with weird "é" still get results #9222
Searching with combined diacritical marks, as can be produced by macOS,
should still match normal "é". To do so, we need to normalized input to
convert the weird one into a normal one. Then the database collation
will actually look for a "e", instead of "é", so it will correctly match
both the normal and weird "é". That means database can be a mix of
normal and weird and searching will work as a human would expect.

In an ideal world the entire database content would be normalized
though. But that's a lot of work...

See https://en.wikipedia.org/wiki/Combining_character