URL are most often optional and thus must accept empty string #9401
In almost all of our use-case the URL type is used for a field that is
optional. Since end-user can only enter an empty string to convey "I
don't have any URL", then we must accept an empty string as a valid URL.

That means DB field must be non-null, and non-unique to accept all
absence of URL as empty strings.

In the very rare cases where an URL is absolutely required, then a
custom type might be implemented in the project.