Search with accent did not return existing result #4638
When a search was done in a localized field it did not always return
results even though they existed in DB. This is because localized
field use JSON to store data in DB and default behavior of Doctrine
is to escape unicode characters. That means a "é" would end up as
"\u00e9". If the search was only "é" that would still correctly matches,
but if the search had a prefix (or suffix) such as "sé", then it would
incorrectly not match.

To fix this, we override Doctrine behavior and always store unescaped
unicode characters. This makes the search straightforward and easier
to work with raw data in DB.