IPs can be configured to bypass signed queries #9696
The configuration `allowedIps` supports a list of IP or subnet in v4 or
v6. If the request is coming from one of those ranges, and it is not
signed at all, then it will be accepted.

However, if the request is signed, the `allowedIps` has no effect at all.
This is because we want to be able to test signed queries even from an
allowed IP.