Felix decides how the client handles errors #9900
Because webonyx/graphql dropped support for `category` we need a new
mechanism to decide how to handle error in the client. This is now the
role of Felix and the client is simplified.

An error may have one the extension set to true:

- `showSnack` means the error has a friendly user message that should be
    shown via a snackbar (not a full page error)
- `objectNotFound` means an object was not found. The full page error
    does not suggest reloading, since it is not an API incompatibility
    issue

All other errors are redirected to full page error with a suggestion to
reload the app because we suspect an incompatibility between client and
server (or a temporary failure of the server, but we don't want to let
the client enter an incoherent state)