Signed queries don't break the request for other middlewares
`$request->getBody()->getContents()` consumes the body without rewinding.
Repeated calls, by other middlewares, will return empty string. So
other middlewares will think the whole request is empty, and GraphQL
will throw.

To avoid this, we should rewind the stream. However, not all streams are
rewindable and would throw exceptions. To avoid that, we return a new
request identical to original with a brand-new stream for body.