Time type complete support #9967
A time in DB uses the `TIME` native MariaDB type, with a precision to
the seconds.

However, the GraphQL time type is meant to be more human friendly and is
only precise to the minute. It also accepts a few formats as input. That
should allow the client code to directly forward whatever the human is
typing.